backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pillow
torch
torchvision
captum
python-multipart
argparse
